requirements.txt
streamlit==1.29.0
pandas==2.0.3
plotly==5.17.0
openai==1.6.1
python-dotenv==1.0.0
google-auth==2.25.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.111.0 